policy/stream-manager: Support routing policy for having priority for the last connected device
Added STREAM_ROUTE_TYPE_AUTO_LAST_CONNECTED route type.
When a stream that has this route type(it should be defined to route-type with "auto-last-connected" in stream-map.json),
it would use a device which is the most recently connected.
Fix mute bug regarding moving sink-inputs.
Do mute/unmute in case of moving source-ouputs.

[Version] 5.0.17
[Profile] Common
[Issue Type] Feature enhancement

Change-Id: I139711b0178ba6f03244735cd728066ff7746c3c
Signed-off-by: Sangchul Lee <sc11.lee@samsung.com>